+++ cmd/tt/subcmd.go
@@ -7,7 +7,6 @@ import (
 	"net"
 	"net/url"
 	"os"
-	"os/signal"
 
 	"github.com/gregoryv/cmdline"
 	"github.com/gregoryv/mq"
@@ -81,13 +80,6 @@ func (c *SubCmd) Run(ctx context.Context) error {
 	p.SetClientID("ttsub")
 	_ = transmit(ctx, p)
 
-	interrupt := make(chan os.Signal, 1)
-	signal.Notify(interrupt, os.Kill, os.Interrupt)
-
-	select {
-	case <-done:
-	case <-interrupt:
-		fmt.Println("interrupted")
-	}
+	<-done
 	return nil
 }